+++ src/fidelichem/gui/views/__init__.py
@@ -1,0 +1,23 @@
+"""GUI view components for FideliChem desktop application."""
+
+from .compounds_view import CompoundsView
+from .decision_view import DecisionView
+from .docking_view import DockingView
+from .dynamics_view import DynamicsView
+from .exports_view import ExportsView
+from .import_view import ImportView
+from .interactions_view import InteractionsView
+from .project_view import ProjectView
+from .qc_view import QCView
+
+__all__ = [
+    "CompoundsView",
+    "DecisionView",
+    "DockingView",
+    "DynamicsView",
+    "ExportsView",
+    "ImportView",
+    "InteractionsView",
+    "ProjectView",
+    "QCView",
+]